Fix parsing error in bytesSum function
This function parses an int array into a hex string by mapping through each integer and converting it to a string. However, sometimes it is passed an Uint8Array, and when mapping that, it autocasts the strings back to numbers, which results in data loss. This fixes it by converting the input to a regular array so we can map through it normally.